Multi-channel augmentation.
This is the initial multi-channel augmentation for pynars v 0.0.2. It is poorly organized since it is under development. Some original pynars class definitions are changed. MultiChannelx.py in the above file are different experiments and small versions. Only MultiChannel6.py and MultiChannel9.py are acknowledged. Run MultiChannel6.py will show the working cycle of one buffer (event buffer). Run MultiChannel9.py will show the entire working cycle, though some functionalities are not implemented as explained in the conceptual design of OpenNARS 3.1.3.
These documents are long and contain low-quality code blocks, but they are the most straightforward implementations, and they are commented very well.