Fetal bugs fixed.
The priority queue for events and predictions are actually from large to small (e.g., 1, 2, 3, ...). But I considered that it is from small to larger, so I used the last one as the candidate.

I have also fixed the bug in making anticipations. In PyNARS, not like OpenNARS, "+1" is allowed in Narsese, so it is possible to have "(&/, A, +1)" as the subject of a prediction. Here I used only "A" as the subject. This time offset is used to put the anticipation in the corresponding time slot. Therefore, I also changed the UI updating method, every time slot will be updated simultaneously.

Enjoy!